requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
cached-property==1.5.2
Django==4.0.3
django-cors-headers==3.11.0
djangorestframework==3.13.1
enum-compat==0.0.3
psycopg2-binary==2.9.3
PyJWT==2.3.0
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
